* Removed entry which was never used in the type array we use to compute the dynamic type of a generic type. This implied changes in the code generation in both frozen and melted mode, as well as updating callers to take into account the removal:  mostly interpreter, reflection (eif_type_id.c) and store/retrieve. This also gets rid of code that were doing a +1 without any explanations. * The above changes has a small impact in size reduction of the generated executable (about 0.5%). * Added some more contracts. * Removed the need for BC_PTUPLE since the generated code is teh same as the one in BC_TUPLE. It frees us an entry for the interpreter byte code. * Updated byte code meltdump and bytedump accordingly. * Fixed some C compiler warnings when comparing signed and unsigned entities (e.g. iregsz in interp.c or using l_body_id in eif_override_byte_code_of_body).
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@93656 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/C/run-time/include/rt_globals.h
@@ -132,7 +132,7 @@ typedef struct tag_rt_globals
 #ifdef WORKBENCH
 	struct opstack op_stack_cx;			/* Operational stack */
 	EIF_TYPED_VALUE **iregs_cx;				/* Interpreter registers */
-	int iregsz_cx;						/* Size of 'iregs' array (bytes) */
+	size_t iregsz_cx;						/* Size of 'iregs' array (bytes) */
 	uint32 argnum_cx;					/* Number of arguments */
 	uint32 locnum_cx;					/* Number of locals */
 	unsigned long tagval_cx;			/* Records number of interpreter's call */
